Add basic CRUD

+++ src/main/java/ba/unsa/etf/ri/ppis/bukvarexpress/entity/OrderEntity.java
@@ -19,8 +19,8 @@ public class OrderEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(columnDefinition = "serial")
     private Long id;
-    private Integer userId;
-    private Integer bookId;
+    private Long userId;
+    private Long bookId;
     private String city;
     private String address;
 }
